chore: move window and fix some warnings
