+++ src/engine/tests/engine/FixedUpdateTest.cpp
@@ -23,7 +23,7 @@ TEST(Core, FixedTimeUpdate)
     // because sleep_for is not precise enough
 
     core.RegisterSystem<Scheduler::FixedTimeUpdate>([&update_count](const Core &) { update_count++; });
-    core.GetScheduler<Scheduler::FixedTimeUpdate>().SetTickRate(1.0 / 5.0);
+    core.GetScheduler<Scheduler::FixedTimeUpdate>().SetTickRate(1.0f / 5.0f);
 
     // First update just after creating should not run the system
     core.RunSystems();
