+++ src/plugin/input/src/Input.pch.hpp
@@ -13,5 +13,4 @@
 
 #include "plugin/PluginWindow.hpp"
 #include "scheduler/Startup.hpp"
-// TODO: put window in resource folder directly rather than putting it in Window subfolder
-#include "resource/Window/Window.hpp"
+#include "resource/Window.hpp"
