+++ src/plugin/window/src/resource/Window.cpp
