+++ src/plugin/window/src/resource/Window.hpp
