+++ src/plugin/window/xmake.lua
@@ -17,7 +17,7 @@ target("PluginWindow")
 
     add_headerfiles("src/(exception/*.hpp)")
     add_headerfiles("src/(plugin/*.hpp)")
-    add_headerfiles("src/(resource/Window/*.hpp)")
+    add_headerfiles("src/(resource/*.hpp)")
     add_headerfiles("src/(system/*.hpp)")
 
     add_includedirs("src", {public = true})
